Refactoring out Reputation-specific data concepts from a general TaskRoll.
This seemed necessary because this roll is not a normal roll where complications and successes are added up.  Reprimand isn't a direct equivalence to a complication, for instance.  They only appear if the roll is failed.
