+++ sta-enhanced.mjs
@@ -1,6 +1,7 @@
 import { STAEnhancedHooks } from './hooks/index.mjs';
+import { TaskRoll } from './dice/TaskRoll.mjs';
 import { ReputationRoll } from './dice/ReputationRoll.mjs';
 
 STAEnhancedHooks.listen();
 
-CONFIG.Dice.rolls.push(ReputationRoll);
+CONFIG.Dice.rolls.push(TaskRoll, ReputationRoll);
