+++ templates/dice/task-tooltip.hbs
